Fix: Card Tabs분기처리 fix #51
