+++ FE/src/Pages/MainPage/MainDish/index.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import * as Carousel from 'Components/Carousel';
-import Card from 'Components/commons/Card.js';
+import Card from 'Components/commons/Cards';
 import API from 'util/API.js';
 
 const MainDish = () => {
